+++ packages/provider-utils/src/index.ts
@@ -5,7 +5,6 @@ export * from './load-api-key';
 export * from './parse-json';
 export * from './post-to-api';
 export * from './response-handler';
-export * from './scale';
 export * from './uint8-utils';
 export * from './validate-types';
 export * from './without-trailing-slash';